+++ CHANGELOG.rst
@@ -6,6 +6,8 @@ Unreleased
 ==========
 
 * Add suport for django CMS 4
+* Fix `468 <https://github.com/django-cms/djangocms-text-ckeditor/issues/468>`_ via `637 <https://github.com/django-cms/djangocms-text-ckeditor/pull/637>`_: Delay importing models.CMSPlugin in utils to allow adding an HTMLField to a custom user model.
+
 
 5.1.1 (2022-06-22)
 ==================